Day5: Part 1 & Part 2 Part 1 was fairly straight forward.
Part 2 was trivial to parse seeds to ranges.  It's computationally complex to manually check each value in the ranges.  Instead, I mapped ranges to their potential destination ranges.  getting the mapping logic down was a little confusing at first, but When you break it down to overlapping number ranges, it becomes easier