Day 18: Part1 and two.
I took the naive approach for part 1 to start, but that wasn't wise.  I found the shoelace formula that I'd used in previous years to calculate the area, but raw area doesn't account for the grid nature of the trenches.

I could have adjusted the points on some sort of scale, but counting the missing perimeter area outside of the center of the blocks did the trick.